ThemeStub: Avoid renaming the default theme
The default theme is default and ThemePicker got their default title inside the app.
By having that, if you put a name that isn't the default or the one that themepicker expects, it makes the app crash due to IllegalArgumentException. In this case it was making the app crash on non-english locales.
So, let's make ThemePicker handle the name and fix the following crash:
msg: java.lang.IllegalArgumentException: Invalid option
stacktrace: java.lang.IllegalArgumentException: Invalid option
	at com.android.customization.widget.OptionSelectorController.setSelectedOption(OptionSelectorController.java:98)
	at com.android.customization.picker.theme.ThemeFragment$3.onOptionsLoaded(ThemeFragment.java:295)
	at com.android.customization.model.theme.DefaultThemeProvider.fetch(DefaultThemeProvider.java:106)
	at com.android.customization.model.theme.ThemeManager.fetchOptions(ThemeManager.java:112)
	at com.android.customization.picker.theme.ThemeFragment.setUpOptions(ThemeFragment.java:265)
	at com.android.customization.picker.theme.ThemeFragment.onViewCreated(ThemeFragment.java:185)