themes: Initial port of our themes for A12
We have a simple color palette to work, so, it won't be 100% accurate but
we can adapt it to the new color palette. Maybe we can improve it later.
Also, transform AOSP Theme into Unmonet just to have fun with the Monet
Deniers (tbh not complaining, monet as an accent is cool and a detox
overlay sounds fun).

Currently the themes have some missing features, like gradients.
Hope I can get time to port it or bring more styles to it.

The following changes are applied:
- Pissel Theme: Because we're already in the pixel style
- OneMono and Layan were kind of boring so, they're gone too
- Some gradients won't be there until I figure out how to bring those back
- Removed custom backgrounds from settings, from the bg shape to the A11 preference styles
- Moved to /product because all overlays should be updated
- Updated transparent themes overlay based on @programminghoch10 mod (ClearLineage)
	(and also, reports from @ouroboros420 that the text was unreadable lol)

Change-Id: Ibf2b2c26aaeb7c4d091abbf33953e5f9eb856e3a