shishufied: packages/overlays/shishufied --> vendor/shishufied [1/3]
This repo was getting huge to be kept as a overlay folder.
So, let's turn it into a vendor for theming and packages we need.

Also, moving the fonts from vendor to here to keep it clean.
Added a toggle for overlays and if we want to ship FontEngine fonts

Change-Id: I95722a9a302e5a18f72dbdeb263eb2237ffcb71b